Analyticsコンポーネントを作成
